requirements_surr_mat.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
ca-certificates=2023.08.22=haa95532_0
certifi=2021.5.30=py36haa95532_0
cycler=0.11.0=pyhd3eb1b0_0
et_xmlfile=1.1.0=py36haa95532_0
freetype=2.12.1=ha860e81_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
intel-openmp=2023.1.0=h59b6b97_46319
jpeg=9e=h2bbff1b_1
julia=0.6.0=pypi_0
kiwisolver=1.3.1=py36hd77b12b_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=1_h8933c1f_netlib
libcblas=3.9.0=5_hd5c7e75_netlib
libdeflate=1.17=h2bbff1b_1
liblapack=3.9.0=5_hd5c7e75_netlib
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
lz4-c=1.9.4=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matlabengineforpython=R2018a=pypi_0
matplotlib=3.3.4=py36haa95532_0
matplotlib-base=3.3.4=py36h49ac443_0
mkl=2020.2=256
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.3.0=py36h46781fe_0
mkl_random=1.1.1=py36h47e9c7a_0
msys2-conda-epoch=20160418=1
numpy=1.19.2=py36hadc3359_0
numpy-base=1.19.2=py36ha3acd2a_0
olefile=0.46=py36_0
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.1.1w=h2bbff1b_0
pandas=1.1.5=py36hd77b12b_0
pillow=8.3.1=py36h4fa10fc_0
pip=21.2.2=py36haa95532_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h6538335_2
python=3.6.13=h3758d61_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.6=2_cp36m
pytz=2021.3=pyhd3eb1b0_0
qt=5.9.7=vc14h73c81de_0
scipy=1.5.2=py36h9439919_0
setuptools=58.0.4=py36haa95532_0
sip=4.19.8=py36h6538335_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
tk=8.6.12=h2bbff1b_0
tornado=6.1=py36h2bbff1b_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py36h7fe50ca_0
xz=5.4.2=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
